Fix building xfsprogs on host systemd with old kernels
The CRC32 selftest compiles natively against the target xfs.h, which
includes <linux/fsmap.h>. On Ubuntu 14.04 (kernel 4.4), fsmap.h doesn't
exist and the build fails.

One workaround would be to disable the selftest, but it could provide
valuable functionality.

To fix this, export STAGING_INCDIR and patch the Makefile to include it
for that one native compilation.

Also delete an unused patch that I missed last time