arch buildhacks: update for master branch of meta-java
The meta-java master branch includes updated versions of OpenJDK as well
as other fixes for building with gcc 8, which means I have a working
openjdk-8-native build now.

A couple -Wno-error flags are still needed as well as the python fixes,
but I can drop the massive ASSUME_PROVIDED hack.