wireguard: don't run depmod during do_install
module.bbclass sets DEPMOD=echo so that the kernel Makefiles don't try
to run depmod, but wireguard's makefile explicitly tries to run 'depmod -a'
which is wrong and often fails.

In wireguard-tools, don't change ${S} and instead add '-C tools' to
EXTRA_OEMAKE so that this patch (and any others in the future) can be
added in wireguard.inc for both module/tools recipes.