Fix openjdk/openjre builds on Linux 5.x
A kernel version check in one of openjdk's Makefiles fails, add the 5.x
series to its list of valid kernel versions.

Only bbappended openjdk-8-native and openjre-8 since that's what I build
for unifi, but more bbappend symlinks could be added if I need to
compile JDK for the target.