java: fix icedtea and openjdk native builds
Another Arch gcc update, another round of breakage.

Add a patch to icedtea to fix C++ complaints about default arguments in
a friend declaration that's not also a definition or something like that.
The patch itself was complicated just to add, since most of the source
is extracted during do_configure and patches are applied separately
through special variables later on.

The meta-java openjdk recipes append compile flags based on major GCC
version, so I need to add a case for GCC9 there.