wireguard: X.0.20191226, rewrite for split repos
As of the 20191226 version, WireGuard is now split into multiple repos
since it's staged to be in the 5.6 mainline release.
wireguard-module-compat is the out-of-tree kernel module source, and
wireguard-tools is the userspace tools.

The versions aren't necessarily in lock-step (0.0.20191226 vs
1.0.20191226) so it doesn't make sense to have wireguard.inc anymore.
While I'm at it, switch to using zx2c4 repos rather than github mirrors,
and save the full SRCREV in the recipe rather than relying on tags.