experiment: use zstd compression for data.tar in IPK packages
I thought this might speed up do_package_write_ipk for large packages,
and maybe do_rootfs for large images, but testing showed that not to be
the case, pigz is still slightly faster.

do_package_write_ipk seems to be very syscall and IO heavy (high kernel
CPU use rather than userland) and compression itself isn't enough of
a bottleneck to make an impact on the overall build time.