add ripgrep recipe, include it in packagegroup-wild-utils
Note: this uses a2x (asciidoc) for manpage generation rather than
asciidoctor. a2x is deprecated in ripgrep, and support for it might go
away some day. There's no asciidoctor recipe in any of my current
layers, and it's a Ruby program so things might get tricky, but that's
a problem for Future Allen to deal with.