fix icedtea7-native build with GCC 12
Mysterious segfaults in java were happening during an "ant build" step
of the build and I wasn't able to find any other bug reports or patches
that seemed relevant.

Downgrading my Arch packages to GCC 11 fixed the build, so it seems that
something in the GCC 12 optimizer is causing a "miscompile" (probably
actually UB in the hotspot jvm code). So I tried building with -O0 and
that worked, so ship it.