langdale: update rust recipes
oe-core now basically includes meta-rust, and meta-rust-bin broke, so
update things to work with the new system. It's like 30+ more minutes of
build time to bootstrap LLVM and rustc, but at least it's cached and
it's nicer to be official. Maybe maybe-rust-bin will get fixed and I can
add it back in.

However, I need some workarounds for the built-in Rust support, mainly
that it's expected that we make recipes for every single dependency
(using cargo-bitbake probably) and then crates.io is ignored. That's
cool and all for reproducibility, but a massive maintenance headache
that I want no part of.

Wrap the common Rust changes I want in cargo-wild.bbclass and use that
form my recipes.