unifi: update to v7.5.187
This major version bump now requires Java 17 rather than Java 8, and
that necessitated a switch from recipes in meta-java to recipes in
meta-openjdk-temurin. Good news is these prebuilt binaries save a ton of
time during builds rather than bootstrapping icedtea and such from
source, which was always a massive headache.

However it's also apparent that we need to pass a lot more options to
java than we used to when launching ace.jar. Write a wrapper script
which injects those options, which were adapted from the unifi.init
script in the official deb packages.