libffi-native: fix build failure
I don't see an upstream fix for this, but I'm struggling to build
master libffi locally for unrelated weird broken automake related
reasons.

This patch is totally a hack, but I'm not sure which header file I'm
supposed to be adding this function prototype to, there's all sorts of
wild preprocessor crap everywhere.