zfs-module: disable separate build directory
ZFS can't compile in a separate build directory on newer kernels, which
I hit when building Linux 6.12.

The problem is described in [1] and fixed in [2] but that fix is already
present in 2.3.4 and the problem still persists in Yocto builds.

I don't know what's wrong, it seems to be an issue deep in the guts of
Kbuild, but building in-tree (${S} == ${B}) seems to work so just do
that.

[1] https://github.com/openzfs/zfs/issues/17321
[2] https://github.com/openzfs/zfs/pull/17517